Create `VerifiedAgeViewModel` to load and hold verification state.
