+++ Samples/Swift/DaysUntilBirthday/Shared/Services/GoogleSignInAuthenticator.swift
@@ -27,32 +27,6 @@ final class GoogleSignInAuthenticator: ObservableObject {
     self.authViewModel = authViewModel
   }
 
-#if os(iOS)
-  /// Verifies the user's age based upon the selected account.
-  /// - note: Successful calls to this will set the `authViewModel`'s `verificationState` property.
-  func verifyAccountDetails() {
-    let accountDetails: [GIDVerifiableAccountDetail] = [
-      GIDVerifiableAccountDetail(accountDetailType:.ageOver18)
-    ]
-
-    guard let rootViewController = UIApplication.shared.windows.first?.rootViewController else {
-      print("There is no root view controller!")
-      return
-    }
-
-    let verifyAccountDetail = GIDVerifyAccountDetail()
-    verifyAccountDetail.verifyAccountDetails(accountDetails, presenting: rootViewController) {
-      verifyResult, error in
-      guard let verifyResult = verifyResult else {
-        self.authViewModel.verificationState = .unverified
-        print("Error! \(String(describing: error))")
-        return
-      }
-      self.authViewModel.verificationState = .verified(verifyResult)
-    }
-  }
-#endif
-
   /// Signs in the user based upon the selected account.'
   /// - note: Successful calls to this will set the `authViewModel`'s `state` property.
   func signIn() {
