+++ Samples/Swift/DaysUntilBirthday/iOS/UserProfileView.swift
@@ -19,6 +19,8 @@ import GoogleSignIn
 
 struct UserProfileView: View {
   @EnvironmentObject var authViewModel: AuthenticationViewModel
+
+  @StateObject var verifiedAgeViewModel = VerifiedAgeViewModel()
   @StateObject var birthdayViewModel = BirthdayViewModel()
   private var user: GIDGoogleUser? {
     return GIDSignIn.sharedInstance.currentUser
@@ -52,9 +54,9 @@ struct UserProfileView: View {
           })
 #if os(iOS)
           NavigationLink(NSLocalizedString("Verify My Age", comment: "Verify Age"),
-                         destination: VerificationView(authViewModel: authViewModel)
+                         destination: VerificationView(verifiedAgeViewModel: verifiedAgeViewModel)
           .onAppear {
-            authViewModel.verifyAccountDetails()
+            verifiedAgeViewModel.verifyAccountDetails()
           })
 #endif
           Spacer()
